requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
Bootstrap-Flask==2.3.0
email_validator==2.1.0
WeasyPrint==60.1
Pillow==10.1.0
